+++ Commands/Base/SPOnlineConnectionHelper.cs
@@ -17,8 +17,8 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using System.Diagnostics;
-using System.Runtime.InteropServices;
 using System.Net.Http;
+using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using Newtonsoft.Json;
 using SharePointPnP.PowerShell.Commands.Utilities;
@@ -28,6 +28,14 @@ namespace SharePointPnP.PowerShell.Commands.Base
 {
     internal class SPOnlineConnectionHelper
     {
+
+#if DEBUG
+        private static readonly Uri VersionCheckUrl = new Uri("https://raw.githubusercontent.com/SharePoint/PnP-PowerShell/dev/version.txt");
+#else
+        private static readonly Uri VersionCheckUrl = new Uri("https://raw.githubusercontent.com/SharePoint/PnP-PowerShell/master/version.txt");
+#endif
+        private static bool VersionChecked;
+
 #if !NETSTANDARD2_0
         public static AuthenticationContext AuthContext { get; set; }
 #endif
@@ -57,7 +65,7 @@ namespace SharePointPnP.PowerShell.Commands.Base
             context.RequestTimeout = requestTimeout;
 #if !ONPREMISES
             context.DisableReturnValueCache = true;
-#elif SP2016
+#elif SP2016 || SP2019
             context.DisableReturnValueCache = true;
 #endif
             var connectionType = ConnectionType.OnPrem;
@@ -97,7 +105,7 @@ namespace SharePointPnP.PowerShell.Commands.Base
         {
             context.ApplicationName = Properties.Resources.ApplicationName;
             context.RequestTimeout = requestTimeout;
-#if SP2016
+#if SP2016 || SP2019
             context.DisableReturnValueCache = true;
 #endif
             var connectionType = ConnectionType.OnPrem;
@@ -378,9 +386,10 @@ namespace SharePointPnP.PowerShell.Commands.Base
             return spoConnection;
         }
 
-        internal static SPOnlineConnection InitiateAzureADAppOnlyConnection(Uri url, string clientId, string tenant, string certificatePEM, string privateKeyPEM, int minimalHealthScore, int retryCount, int retryWait, int requestTimeout, string tenantAdminUrl, PSHost host, bool disableTelemetry, bool skipAdminCheck = false, AzureEnvironment azureEnvironment = AzureEnvironment.Production)
+        internal static SPOnlineConnection InitiateAzureADAppOnlyConnection(Uri url, string clientId, string tenant, string certificatePEM, string privateKeyPEM, SecureString certificatePassword, int minimalHealthScore, int retryCount, int retryWait, int requestTimeout, string tenantAdminUrl, PSHost host, bool disableTelemetry, bool skipAdminCheck = false, AzureEnvironment azureEnvironment = AzureEnvironment.Production)
         {
-            X509Certificate2 certificate = CertificateHelper.GetCertificateFromPEMstring(certificatePEM, privateKeyPEM);
+            string password = new System.Net.NetworkCredential(string.Empty, certificatePassword).Password;
+            X509Certificate2 certificate = CertificateHelper.GetCertificateFromPEMstring(certificatePEM, privateKeyPEM, password);
 
             var authManager = new OfficeDevPnP.Core.AuthenticationManager();
             var clientContext = authManager.GetAzureADAppOnlyAuthenticatedContext(url.ToString(), clientId, tenant, certificate, azureEnvironment);
@@ -397,7 +406,27 @@ namespace SharePointPnP.PowerShell.Commands.Base
                     connectionType = ConnectionType.TenantAdmin;
                 }
             }
+
+            CleanupCryptoMachineKey(certificate);
+
             return new SPOnlineConnection(context, connectionType, minimalHealthScore, retryCount, retryWait, null, url.ToString(), tenantAdminUrl, PnPPSVersionTag, host, disableTelemetry);
+        }
+
+        private static void CleanupCryptoMachineKey(X509Certificate2 certificate)
+        {
+            var privateKey = (RSACryptoServiceProvider) certificate.PrivateKey;
+            string uniqueKeyContainerName = privateKey.CspKeyContainerInfo.UniqueKeyContainerName;
+            certificate.Reset();
+            var path = Environment.GetEnvironmentVariable("ProgramData");
+            if (string.IsNullOrEmpty(path)) path = @"C:\ProgramData";
+            try
+            {
+                System.IO.File.Delete(string.Format(@"{0}\Microsoft\Crypto\RSA\MachineKeys\{1}", path, uniqueKeyContainerName));
+            }
+            catch (Exception)
+            {
+                // best effort cleanup
+            }
         }
 #endif
 #endif
@@ -435,7 +464,7 @@ namespace SharePointPnP.PowerShell.Commands.Base
                 context.RequestTimeout = requestTimeout;
 #if !ONPREMISES
                 context.DisableReturnValueCache = true;
-#elif SP2016
+#elif SP2016 || SP2019
             context.DisableReturnValueCache = true;
 #endif
                 var connectionType = ConnectionType.OnPrem;
@@ -467,7 +496,7 @@ namespace SharePointPnP.PowerShell.Commands.Base
             context.ApplicationName = Properties.Resources.ApplicationName;
 #if !ONPREMISES
             context.DisableReturnValueCache = true;
-#elif SP2016
+#elif SP2016 || SP2019
             context.DisableReturnValueCache = true;
 #endif
             context.RequestTimeout = requestTimeout;
@@ -567,7 +596,7 @@ namespace SharePointPnP.PowerShell.Commands.Base
             context.RequestTimeout = requestTimeout;
 #if !ONPREMISES
             context.DisableReturnValueCache = true;
-#elif SP2016
+#elif SP2016 || SP2019
             context.DisableReturnValueCache = true;
 #endif
 
@@ -692,5 +721,34 @@ namespace SharePointPnP.PowerShell.Commands.Base
                 return (result);
             },
             true);
+
+        public static string GetLatestVersion()
+        {
+            try
+            {
+                if (!VersionChecked)
+                {
+                    using (var client = new WebClient())
+                    {
+                        SPOnlineConnectionHelper.VersionChecked = true;
+                        var onlineVersion = client.DownloadString(VersionCheckUrl);
+                        onlineVersion = onlineVersion.Trim(new char[] { '\t', '\r', '\n' });
+                        var assembly = Assembly.GetExecutingAssembly();
+                        var currentVersion = ((AssemblyFileVersionAttribute)assembly.GetCustomAttribute(typeof(AssemblyFileVersionAttribute))).Version;
+                        if (Version.TryParse(onlineVersion, out Version availableVersion))
+                        {
+                            if (availableVersion > new Version(currentVersion))
+                            {
+                                return $"A newer version of PnP PowerShell is available: {availableVersion}. Consider upgrading.";
+                            }
+                        }
+                        SPOnlineConnectionHelper.VersionChecked = true;
+                    }
+                }
+            }
+            catch
+            { }
+            return null;
+        }
     }
 }