+++ Commands/ClientSidePages/AddClientSidePageSection.cs
@@ -23,6 +23,11 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
 PS> Add-PnPClientSidePageSection -Page $page -SectionTemplate OneColumn",
         Remarks = "Adds a new one column section to the Client-Side page 'MyPage'",
         SortOrder = 3)]
+    [CmdletExample(
+        Code = @"PS:> $page = Add-PnPClientSidePage -Name ""MyPage""
+PS> Add-PnPClientSidePageSection -Page $page -SectionTemplate OneColumn -ZoneEmphasis 2",
+        Remarks = "Adds a new one column section to the Client-Side page 'MyPage' and sets the background to 2 (0 is no background, 3 is highest emphasis)",
+        SortOrder = 4)]
     public class AddClientSidePageSection : PnPWebCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0, HelpMessage = "The name of the page")]
@@ -34,6 +39,9 @@ PS> Add-PnPClientSidePageSection -Page $page -SectionTemplate OneColumn",
         [Parameter(Mandatory = false, HelpMessage = "Sets the order of the section. (Default = 1)")]
         public int Order = 1;
 
+        [Parameter(Mandatory = false, HelpMessage = "Sets the background of the section (default = 0)")]
+        public int ZoneEmphasis = 0;
+
 
         protected override void ExecuteCmdlet()
         {
@@ -41,7 +49,7 @@ PS> Add-PnPClientSidePageSection -Page $page -SectionTemplate OneColumn",
 
             if (clientSidePage != null)
             {
-                clientSidePage.AddSection(SectionTemplate, Order);
+                clientSidePage.AddSection(new CanvasSection(clientSidePage, SectionTemplate, Order) { ZoneEmphasis = ZoneEmphasis });
                 clientSidePage.Save();
             }
             else